requirements.txt
flask
gunicorn
requests
mysql-connector-python